gabc corpora: import: use lygre's SimpleGabcParser
which does not attempt to parse music and therefore manages to process
many scores which the stricter/more detailed GabcParser cannot.

Since lygre is old, was meant as a CLI tool rather than a library and
does not have its constants wrapped in a module (and this app follows
default Rails app naming conventions and does not wrap its constants in
a module either) there was a name clash between our application code and
a class recently added to lygre. Scary.