Volpiano: from MEI: start custom processing
as the pure music21 way doesn't provide satisfactory results.

Unfortunately the loading via music21.mei seems to be losing some
lyrics-related information in the loading process, which prevents us
from correctly restructuring the data according to the lyrics structure,
but the results are already good enough for the most common music search
scenarios.