delete remains of "const_attributes"
it was a poor design choice meant to save typing and eliminate repeated
code, but saving very little if anything at all.

The actual solution was not to pass attribute values common to the whole
import or to a large portion of the imported data through the Adapter
and to modify the (un)implemented attributes reporting to account for
these attributes.